refactor: split remaining napi modules
